refactor: move admin components to admin folder
